Amend to accommodate for html & css change
Rearranged html so the buttons can be replaced to the restart btn.
Probably not the most efficient way, but by putting the reset button in
a container, I was able to replace the two containers at the end of the
game, so had to amend the reset event listeners too. There might be an easier option where I don't have to iterate through that one element inside
it, but for now, I couldn't find it.