Fix visibility, btn containers, events
Use containers for visibility, hide/unhide

Wanted to use display: block but block doesn't work with flex so changed it to visibility. Like this the boxes don't get swapped

Fixed the issue of having to iterate over the contents of the button containers while also taking away the reset button container as for once it is not needed and for second it was in the way and did not allow to add height to it for some reason.

Also made elements hide and show up at different events.

Added extra messages that show up for certain events

Changed text color in many places depending on the "result"